chore(core): backend restructuration (8)
