+++ src/bp/bpfs_recovery.ts
@@ -5,14 +5,14 @@ import './sdk/rewire'
 import './common/polyfills'
 
 import chalk from 'chalk'
+import { container } from 'core/app/app.inversify'
+import { GhostService } from 'core/bpfs'
+import Database from 'core/database'
+import { TYPES } from 'core/types'
 import fs from 'fs'
 import { mkdirpSync } from 'fs-extra'
 import path from 'path'
 import 'reflect-metadata'
-import { container } from './core/app.inversify'
-import Database from './core/database'
-import { GhostService } from './core/services'
-import { TYPES } from './core/types'
 
 export default async (argv, action) => {
   let ghost: GhostService | undefined
