+++ src/bp/core/api.ts
@@ -1,4 +1,5 @@
 import * as sdk from 'botpress/sdk'
+import { container } from 'core/app/app.inversify'
 import { BotService } from 'core/bots'
 import { GhostService } from 'core/bpfs'
 import { CMSService, renderRecursive, RenderService } from 'core/cms'
@@ -17,6 +18,7 @@ import { NotificationsService } from 'core/notifications'
 import { WellKnownFlags } from 'core/sdk/enums'
 import { getMessageSignature } from 'core/security'
 import { TYPES } from 'core/types'
+import { HookService } from 'core/user-code'
 import { ChannelUserRepository, WorkspaceService } from 'core/users'
 import { inject, injectable } from 'inversify'
 import Knex from 'knex'
@@ -24,10 +26,8 @@ import _ from 'lodash'
 import { Memoize } from 'lodash-decorators'
 import MLToolkit from 'nlu/ml/toolkit'
 
-import { container } from './app.inversify'
 import { Event, RealTimePayload } from './sdk/impl'
 import HTTPServer from './server'
-import { HookService } from './services/hook/hook-service'
 import RealtimeService from './services/realtime'
 
 const http = (httpServer: HTTPServer) => (identity: string): typeof sdk.http => {
