+++ src/bp/core/app.ts
@@ -1,13 +1,13 @@
 import 'bluebird-global'
 import LicensingService from 'common/licensing-service'
+import { ConfigProvider } from 'core/config'
 import { LoggerProvider } from 'core/logger'
-import { LocalActionServer as LocalActionServerImpl } from 'core/services/action/local-action-server'
+import { LocalActionServer as LocalActionServerImpl } from 'core/user-code'
 import { FatalError } from 'errors'
 import 'reflect-metadata'
 
-import { container } from './app.inversify'
+import { container } from './app/app.inversify'
 import { Botpress as Core } from './botpress'
-import { ConfigProvider } from './config/config-loader'
 import Database from './database'
 import { GhostService } from './services'
 import { TYPES } from './types'
