+++ src/bp/core/app/app.inversify.ts
@@ -1,24 +1,23 @@
 import { Logger } from 'botpress/sdk'
+import { BotpressAPIProvider } from 'core/api'
+import { Botpress } from 'core/botpress'
 import { ConfigProvider } from 'core/config'
 import { EventCollector } from 'core/events'
+import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, PersistedConsoleLogger } from 'core/logger'
+import { MigrationService } from 'core/migration'
 import { ModuleLoader } from 'core/modules'
+import HTTPServer from 'core/server'
 import { TelemetryContainerModules } from 'core/telemetry'
 import { TYPES } from 'core/types'
-import { DataRetentionJanitor, DataRetentionService } from 'core/users'
+import { LocalActionServer } from 'core/user-code'
+import { DataRetentionJanitor, DataRetentionService, WorkspaceService } from 'core/users'
 import { Container } from 'inversify'
 
-import { BotpressAPIProvider } from './api'
-import { Botpress } from './botpress'
-import { DatabaseContainerModules } from './database/database.inversify'
-import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, PersistedConsoleLogger } from './logger'
-import { applyDisposeOnExit, applyInitializeFromConfig } from './misc/inversify'
-import { RepositoriesContainerModules } from './repositories/repositories.inversify'
-import HTTPServer from './server'
-import { LocalActionServer } from './services/action/local-action-server'
-import { MigrationService } from './services/migration'
-import { ServicesContainerModules } from './services/services.inversify'
-import { WorkspaceService } from './services/workspace-service'
-import { Statistics } from './stats'
+import { applyDisposeOnExit, applyInitializeFromConfig } from '../misc/inversify'
+import { Statistics } from '../stats'
+import { DatabaseContainerModules } from './database.inversify'
+import { RepositoriesContainerModules } from './repositories.inversify'
+import { ServicesContainerModules } from './services.inversify'
 
 const container = new Container({ autoBindInjectable: true })
