+++ src/bp/core/app/database.inversify.ts
@@ -1,9 +1,8 @@
+import Database from 'core/database'
+import { patchKnex } from 'core/database/helpers'
 import { TYPES } from 'core/types'
 import { ContainerModule, interfaces } from 'inversify'
 import Knex from 'knex'
-
-import Database from '.'
-import { patchKnex } from './helpers'
 
 const DatabaseContainerModule = new ContainerModule((bind: interfaces.Bind) => {
   bind<Database>(TYPES.Database)
