+++ src/bp/core/app/services.inversify.ts
@@ -7,6 +7,7 @@ import { ConverseService } from 'core/converse'
 import { DialogContainerModule, SkillService } from 'core/dialog'
 import { CEJobService, JobService } from 'core/distributed'
 import { EventEngine, Queue, MemoryQueue } from 'core/events'
+import { CEMonitoringService, MonitoringService, AlertingService, CEAlertingService } from 'core/health'
 import { KeyValueStore } from 'core/kvs'
 import { LogsJanitor, LogsService } from 'core/logger'
 import { MediaServiceProvider } from 'core/media'
@@ -15,17 +16,13 @@ import { NotificationsService } from 'core/notifications'
 import { AuthService, AuthStrategies, CEAuthStrategies } from 'core/security'
 import { StatsService } from 'core/telemetry'
 import { TYPES } from 'core/types'
+import { HookService, ActionService, ActionServersService } from 'core/user-code'
 import { ContainerModule, interfaces } from 'inversify'
 
-import ActionServersService from './action/action-servers-service'
-import ActionService from './action/action-service'
-import { AlertingService, CEAlertingService } from './alerting-service'
-import { HintsService } from './hints'
-import { HookService } from './hook/hook-service'
-import CELicensingService from './licensing'
-import { CEMonitoringService, MonitoringService } from './monitoring'
-import { NLUService } from './nlu/nlu-service'
-import RealtimeService from './realtime'
+import { HintsService } from '../services/hints'
+import CELicensingService from '../services/licensing'
+import { NLUService } from '../services/nlu/nlu-service'
+import RealtimeService from '../services/realtime'
 
 const ServicesContainerModule = new ContainerModule((bind: interfaces.Bind) => {
   bind<ConversationService>(TYPES.ConversationService)
