+++ src/bp/core/botpress.ts
@@ -10,13 +10,16 @@ import Database from 'core/database'
 import { StateManager, DecisionEngine, DialogEngine, DialogJanitor } from 'core/dialog'
 import { SessionIdFactory } from 'core/dialog/sessions'
 import { addStepToEvent, EventCollector, StepScopes, StepStatus, EventEngine } from 'core/events'
+import { AlertingService, MonitoringService } from 'core/health'
 import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } from 'core/logger'
+import { MigrationService } from 'core/migration'
 import { copyDir } from 'core/misc/pkg-fs'
 import { ModuleLoader } from 'core/modules'
 import { NotificationsService } from 'core/notifications'
 import { AuthService } from 'core/security'
 import { StatsService } from 'core/telemetry'
 import { TYPES } from 'core/types'
+import { ActionServersConfigSchema, Hooks, HookService } from 'core/user-code'
 import { DataRetentionJanitor, DataRetentionService, WorkspaceService } from 'core/users'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
@@ -35,12 +38,7 @@ import { setDebugScopes } from '../debug'
 import { WellKnownFlags } from './sdk/enums'
 import { Event } from './sdk/impl'
 import HTTPServer from './server'
-import { ActionServersConfigSchema } from './services/action/action-servers-service'
-import { AlertingService } from './services/alerting-service'
 import { HintsService } from './services/hints'
-import { Hooks, HookService } from './services/hook/hook-service'
-import { MigrationService } from './services/migration'
-import { MonitoringService } from './services/monitoring'
 import RealtimeService from './services/realtime'
 import { Statistics } from './stats'
