+++ src/bp/core/bpfs/index.ts
@@ -3,3 +3,4 @@
 export * from '../services/ghost'
 export * from '../services/ghost/service'
 export * from '../services/ghost/ghost.inversify'
+export * from '../services/ghost/index'
