--- src/bp/core/database/tables/bot-specific/index.ts
@@ -1,1 +1,0 @@
-export * from './tasks'
