--- src/bp/core/database/tables/server-wide/index.ts
@@ -1,2 +1,0 @@
-export * from './metadata'
-export * from './migrations'
