+++ src/bp/core/dialog/index.ts
@@ -7,3 +7,4 @@ export * from '../services/dialog/janitor'
 export * from '../services/dialog/flow/service'
 export * from '../services/dialog/skill/service'
 export * from '../services/dialog/dialog.inversify'
+export * from '../services/dialog/errors'
