+++ src/bp/core/events/event-engine.ts
@@ -1,7 +1,8 @@
 import * as sdk from 'botpress/sdk'
+import { incrementMetric } from 'core/health'
 import { TimedPerfCounter } from 'core/misc/timed-perf'
 import { WellKnownFlags } from 'core/sdk/enums'
-import { incrementMetric } from 'core/services/monitoring'
+import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
 import joi from 'joi'
 import _ from 'lodash'
@@ -9,8 +10,6 @@ import { VError } from 'verror'
 import yn from 'yn'
 
 import { Event } from '../sdk/impl'
-import { TYPES } from '../types'
-
 import { EventCollector } from './event-collector'
 import { MiddlewareChain } from './middleware-chain'
 import { Queue } from './queue'
