+++ src/bp/core/health/index.ts
@@ -1,0 +1,4 @@
+// TODO: Move code here when refactoring is complete
+
+export * from '../services/alerting-service'
+export * from '../services/monitoring'
