+++ src/bp/core/logger/logs-janitor.ts
@@ -1,7 +1,6 @@
 import { Logger } from 'botpress/sdk'
-import { BotpressConfig } from 'core/config/botpress.config'
-import { ConfigProvider } from 'core/config/config-loader'
-import { BotService } from 'core/services/bot-service'
+import { BotService } from 'core/bots'
+import { BotpressConfig, ConfigProvider } from 'core/config'
 import { Janitor } from 'core/services/janitor'
 import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
