+++ src/bp/core/logger/persister/console-logger.ts
@@ -2,10 +2,10 @@ import { AxiosError } from 'axios'
 import { IO, Logger, LoggerEntry, LoggerLevel, LoggerListener, LogLevel } from 'botpress/sdk'
 import chalk from 'chalk'
 import { Metric } from 'common/monitoring'
+import { BotService } from 'core/bots'
 import { addLogToEvent } from 'core/events'
+import { incrementMetric } from 'core/health'
 import { IDisposable } from 'core/misc/disposable'
-import { BotService } from 'core/services/bot-service'
-import { incrementMetric } from 'core/services/monitoring'
 import { TYPES } from 'core/types'
 import { InvalidParameterError } from 'errors'
 import { EventEmitter2 } from 'eventemitter2'
