+++ src/bp/core/messaging/message-service.ts
@@ -1,9 +1,9 @@
 import * as sdk from 'botpress/sdk'
+import { JobService } from 'core/distributed'
 import { EventEngine } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { MessageRepository } from 'core/messaging'
 import { IOEvent } from 'core/sdk/impl'
-import { JobService } from 'core/services/job-service'
 import { TYPES } from 'core/types'
 import { inject, injectable, postConstruct } from 'inversify'
 import { AppLifecycle, AppLifecycleEvents } from 'lifecycle'
