+++ src/bp/core/migration/index.ts
@@ -1,0 +1,4 @@
+export * from './bot-migration-service'
+
+// TODO: Move code here when references are updated
+export * from '../services/migration'
