+++ src/bp/core/realtime/index.ts
@@ -1,0 +1,3 @@
+// TODO: Move code here when references are updated
+
+export * from '../services/realtime'
