+++ src/bp/core/routers/bots/index.ts
@@ -19,9 +19,8 @@ import {
   fileUploadMulter,
   needPermissions
 } from 'core/security'
-import ActionServersService from 'core/services/action/action-servers-service'
-import ActionService from 'core/services/action/action-service'
 import { getSocketTransports } from 'core/services/realtime'
+import { ActionService, ActionServersService } from 'core/user-code'
 import { WorkspaceService } from 'core/users'
 import { Express, RequestHandler, Router } from 'express'
 import { validate } from 'joi'
