+++ src/bp/core/server.ts
@@ -13,6 +13,7 @@ import { ExternalAuthConfig, ConfigProvider } from 'core/config'
 import { ConverseRouter, ConverseService } from 'core/converse'
 import { FlowService, SkillService } from 'core/dialog'
 import { JobService } from 'core/distributed'
+import { AlertingService, MonitoringService } from 'core/health'
 import { LogsService, LogsRepository } from 'core/logger'
 import { MediaServiceProvider, MediaRouter } from 'core/media'
 import { ModuleLoader, ModulesRouter } from 'core/modules'
@@ -30,6 +31,7 @@ import {
 } from 'core/security'
 import { TelemetryRouter, TelemetryRepository } from 'core/telemetry'
 import { TYPES } from 'core/types'
+import { ActionService, ActionServersService } from 'core/user-code'
 import { WorkspaceService } from 'core/users'
 import cors from 'cors'
 import errorHandler from 'errorhandler'
@@ -59,11 +61,7 @@ import { isDisabled } from './routers/conditionalMiddleware'
 import { InvalidExternalToken, PaymentRequiredError } from './routers/errors'
 import { SdkApiRouter } from './routers/sdk/router'
 import { ShortLinksRouter } from './routers/shortlinks'
-import ActionServersService from './services/action/action-servers-service'
-import ActionService from './services/action/action-service'
-import { AlertingService } from './services/alerting-service'
 import { HintsService } from './services/hints'
-import { MonitoringService } from './services/monitoring'
 import { NLUService } from './services/nlu/nlu-service'
 
 const BASE_API_PATH = '/api/v1'
