+++ src/bp/core/services/bot-service.ts
@@ -3,9 +3,12 @@ import cluster from 'cluster'
 import { BotHealth, ServerHealth } from 'common/typings'
 import { BotCreationSchema, BotEditSchema, isValidBotId } from 'common/validation'
 import { createForGlobalHooks } from 'core/api'
+import { FileContent, GhostService, ReplaceContent } from 'core/bpfs'
 import { CMSService } from 'core/cms'
 import { ConfigProvider } from 'core/config'
+import { JobService } from 'core/distributed'
 import { PersistedConsoleLogger } from 'core/logger'
+import { MigrationService } from 'core/migration'
 import { IDisposable } from 'core/misc/disposable'
 import { listDir } from 'core/misc/list-dir'
 import { stringify } from 'core/misc/utils'
@@ -13,6 +16,8 @@ import { ModuleResourceLoader, ModuleLoader } from 'core/modules'
 import { RealTimePayload } from 'core/sdk/impl'
 import { Statistics } from 'core/stats'
 import { TYPES } from 'core/types'
+import { Hooks, HookService } from 'core/user-code'
+import { WorkspaceService } from 'core/users'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
 import glob from 'glob'
@@ -30,13 +35,8 @@ import { VError } from 'verror'
 import { extractArchive } from '../misc/archive'
 
 import { InvalidOperationError } from './auth/errors'
-import { ReplaceContent } from './ghost'
-import { FileContent, GhostService } from './ghost/service'
-import { Hooks, HookService } from './hook/hook-service'
-import { JobService } from './job-service'
-import { MigrationService } from './migration'
+
 import RealtimeService from './realtime'
-import { WorkspaceService } from './workspace-service'
 
 const BOT_DIRECTORIES = ['actions', 'flows', 'entities', 'content-elements', 'intents', 'qna']
 const BOT_CONFIG_FILENAME = 'bot.config.json'
