+++ src/bp/core/services/dialog/dialog-engine.ts
@@ -2,14 +2,14 @@ import { FlowNode, IO, Logger } from 'botpress/sdk'
 import { FlowView } from 'common/typings'
 import { createForGlobalHooks } from 'core/api'
 import { buildUserKey, converseApiEvents } from 'core/converse'
+import { FlowService } from 'core/dialog'
 import { addErrorToEvent } from 'core/events'
 import { TYPES } from 'core/types'
+import { Hooks, HookService } from 'core/user-code'
 import { inject, injectable, tagged } from 'inversify'
 import _ from 'lodash'
 
-import { Hooks, HookService } from '../hook/hook-service'
 import { FlowError, TimeoutNodeNotFound } from './errors'
-import { FlowService } from './flow/service'
 import { Instruction } from './instruction'
 import { InstructionProcessor } from './instruction/processor'
 import { InstructionQueue } from './instruction/queue'
