+++ src/bp/core/services/migration/index.ts
@@ -1,9 +1,11 @@
 import * as sdk from 'botpress/sdk'
 import chalk from 'chalk'
 import { BotpressAPIProvider } from 'core/api'
-import { ConfigProvider } from 'core/config/config-loader'
+import { GhostService } from 'core/bpfs'
+import { ConfigProvider } from 'core/config'
 import Database from 'core/database'
 import { PersistedConsoleLogger, centerText } from 'core/logger'
+import { BotMigrationService } from 'core/migration'
 import { TYPES } from 'core/types'
 import fse from 'fs-extra'
 import glob from 'glob'
@@ -14,9 +16,7 @@ import semver from 'semver'
 import stripAnsi from 'strip-ansi'
 import yn from 'yn'
 
-import { container } from '../../app.inversify'
-import { GhostService } from '../ghost/service'
-import { BotMigrationService } from './bot'
+import { container } from '../../app/app.inversify'
 
 const debug = DEBUG('migration')
